Extract REM team routing modules
