+++ crates/r3akt-profile-rch/src/fields.rs
@@ -1,0 +1,4 @@
+pub const FIELD_COMMANDS: i64 = 0x09;
+pub const FIELD_RESULTS: i64 = 0x0A;
+pub const FIELD_GROUP: i64 = 0x0B;
+pub const FIELD_EVENT: i64 = 0x0D;
