+++ crates/r3akt-rch-core/src/lib.rs
@@ -37,6 +37,7 @@ use text::decode_utf8_ignoring_errors;
 
 pub mod python_migration;
 mod rem_team_directory;
+mod rem_team_scope;
 
 pub const DELIVERY_ENVELOPE_FIELD: &str = "RTHDelivery";
 pub const DELIVERY_SCHEMA_VERSION: &str = "1";
@@ -4554,12 +4555,8 @@ impl RchCore {
                 format!("Unsupported mission command '{}'", command.command_type),
             ))];
         }
-        if let Err((reason_code, reason)) = self.validate_rem_team_scope(command) {
-            return vec![MissionSyncResponse::results(Self::rejected_result(
-                command,
-                reason_code,
-                reason,
-            ))];
+        if let Some(rejection) = self.rem_team_scope_rejection(command) {
+            return vec![rejection];
         }
         if let Some(required_capability) = required_capability(command.command_type.as_str()) {
             if self.authorization_required
@@ -4679,12 +4676,8 @@ impl RchCore {
                 format!("Unsupported checklist command '{}'", command.command_type),
             ))];
         }
-        if let Err((reason_code, reason)) = self.validate_rem_team_scope(command) {
-            return vec![MissionSyncResponse::results(Self::rejected_result(
-                command,
-                reason_code,
-                reason,
-            ))];
+        if let Some(rejection) = self.rem_team_scope_rejection(command) {
+            return vec![rejection];
         }
         if let Some(required_capability) =
             checklist_required_capability(command.command_type.as_str())
@@ -4752,37 +4745,6 @@ impl RchCore {
         reason: impl Into<String>,
     ) -> CommandResultEnvelope {
         Self::rejected_result_with_capabilities(command, reason_code, reason, Vec::new())
-    }
-
-    fn validate_rem_team_scope(
-        &self,
-        command: &MissionCommandEnvelope,
-    ) -> Result<(), (&'static str, String)> {
-        let Some(team_uid) = command
-            .args
-            .get("_rem_team_uid")
-            .and_then(Value::as_str)
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-        else {
-            return Ok(());
-        };
-        if canonical_team_for_uid(team_uid).is_none() {
-            return Err((
-                "invalid_team",
-                "FIELD_GROUP must contain a canonical TEAM UID".to_string(),
-            ));
-        }
-        if !self
-            .shared_team_uids_for_rem_source(&command.source.rns_identity)
-            .contains(team_uid)
-        {
-            return Err((
-                "unauthorized_team",
-                "The REM caller is not a member of the requested TEAM".to_string(),
-            ));
-        }
-        Ok(())
     }
 
     fn rejected_result_with_capabilities(
